requirements.txt
discord.py[voice]
asyncpg
wavelink
Pillow
psutil
async-timeout
humanize
uvloop; sys_platform == "linux"